Cleanup @emotion/jest entrypoints
