+++ packages/jest/enzyme-serializer/index.js
@@ -1,0 +1,1 @@
+export * from '../src/create-enzyme-serializer'
