+++ packages/jest/enzyme-serializer/package.json
@@ -1,8 +1,8 @@
 {
   "main": "dist/jest.cjs.js",
   "module": "dist/jest.esm.js",
-  "types": "../types/enzyme",
+  "types": "../types/enzyme-serializer",
   "preconstruct": {
-    "source": "../src/enzyme"
+    "source": "../src/enzyme-serializer"
   }
 }
