--- packages/jest/enzyme/index.js
@@ -1,1 +1,0 @@
-export * from '../src/enzyme'
