--- packages/jest/serializer.js
@@ -1,1 +1,0 @@
-module.exports = require('jest-emotion').default
