+++ packages/jest/serializer/index.js
@@ -1,0 +1,1 @@
+export * from '../src/create-serializer'
