+++ packages/jest/serializer/package.json
@@ -1,0 +1,8 @@
+{
+  "main": "dist/jest.cjs.js",
+  "module": "dist/jest.esm.js",
+  "types": "../types/serializer",
+  "preconstruct": {
+    "source": "../src/serializer"
+  }
+}
