+++ packages/jest/src/enzyme-serializer.js
@@ -1,0 +1,3 @@
+// @flow
+import { createEnzymeSerializer } from './create-enzyme-serializer'
+export const { test, serialize } = createEnzymeSerializer()
