+++ packages/jest/src/index.js
@@ -1,3 +1,4 @@
 // @flow
-export { createSerializer, default } from './serializer'
+export { createSerializer } from './create-serializer'
+export { createEnzymeSerializer } from './create-enzyme-serializer'
 export { matchers } from './matchers'
