+++ packages/jest/test/preact.test.js
@@ -5,8 +5,10 @@ import { h } from 'preact'
 import render from 'preact-render-to-json'
 import prettyFormat from 'pretty-format'
 import * as emotion from '@emotion/css'
-import emotionPlugin from '@emotion/jest'
+import { createSerializer } from '@emotion/jest'
 const { ReactElement, ReactTestComponent, DOMElement } = prettyFormat.plugins
+
+let emotionPlugin = createSerializer()
 
 describe('jest-emotion with preact', () => {
   const divStyle = emotion.css`
