+++ packages/jest/types/enzyme-serializer.d.ts
@@ -1,0 +1,9 @@
+/// <reference types="jest" />
+
+type SnapshotSerializerPlugin = Extract<
+  jest.SnapshotSerializerPlugin,
+  { serialize: any }
+>
+
+export const test: SnapshotSerializerPlugin['test']
+export const serialize: SnapshotSerializerPlugin['serialize']
