+++ packages/jest/types/tslint.json
@@ -1,10 +1,8 @@
 {
-    "extends": "dtslint/dtslint.json",
-    "rules": {
-        "array-type": [
-            true,
-            "generic"
-        ],
-        "semicolon": false
-    }
+  "extends": "dtslint/dtslint.json",
+  "rules": {
+    "array-type": [true, "generic"],
+    "semicolon": false,
+    "file-name-casing": false
+  }
 }
