+++ scripts/test-utils/legacy-env.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-env jest */
 import 'test-utils/enzyme-env'
-import serializer from '@emotion/jest/enzyme'
+import { createEnzymeSerializer } from '@emotion/jest'
 
 // $FlowFixMe jest flow type definitions don't include new plugin API
-expect.addSnapshotSerializer(serializer)
+expect.addSnapshotSerializer(createEnzymeSerializer())
